% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baseFlow_DG.R
\name{baseFlow_DG}
\alias{baseFlow_DG}
\title{NOTE - THE SAME AS 'BASEFLOW.R' EXCEPT FOR THAT THIS CALCULATES BFI TOO (USED IN 'EVENTBASEFLOW.R') - TO CHOOSE A VERSION AND COMBINE
#############################
Baseflow removal (after Ladson et al)}
\usage{
baseFlow_DG(q, alpha = 0.925, passes = 3, r = 30)
}
\arguments{
\item{q}{The vector series of streamflow}

\item{alpha}{Filter parameter}

\item{passes}{Number of passes}

\item{r}{number of points refelcted at start and end of data set}
}
\value{
Function returns a vector represeting the baseflow of into the inflow \code{q}
}
\description{
This function removes baseflow based on a recursive digital filter and is based on
the implementation described in Ladson et al (2013).
}
\details{
The reflected points act to resolve spin up issues and are removed before the baseflow is removed.
}
\examples{
data("bassRiver")
bf = baseFlow(bassRiver, alpha = 0.98)
BFI = sum(bf)/sum(bassRiver)
print(BFI) # 0.20
plot(1:length(bassRiver), bassRiver, type = "l", lwd = 2, col = "steelblue",
 ylab = "Flow (ML/d)", xlab = "Time index", mgp = c(2, 0.6, 0))
lines(1:length(bassRiver), bf, lwd = 2, lty = 2, col = "darkgreen")
legend("topright", legend = c("Flow", "Baseflow"), cex = 0.8,
 lwd = 2, col = c("steelblue", "darkgreen"), bty = "n")
}
\references{
Ladson, A., Brown, R., Neal, B., & Nathan, R. (2013). A standard approach to baseflow separation using the Lyne and Hollick filter.
Australian Journal of Water Resources, 17(1).
}
\keyword{baseflow}
