% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cats.R
\name{cats}
\alias{cats}
\title{A Cat Function}
\usage{
cats(love = TRUE)
}
\arguments{
\item{love}{Do you love cats? Defaults to TRUE.}
}
\description{
This function allows one to express their love of cats
}
\examples{
cats()
}
\keyword{cats}
