% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data232210}
\alias{data232210}
\title{Water quality data}
\format{
blah blah

blah blah
}
\source{
\url{https://data.water.vic.gov.au/}

\url{http://www.bom.gov.au/water/hrs/#id=314213/}
}
\usage{
data232210

data232210
}
\description{
Example water quality data for MOORABOOL RIVER WEST BRANCH at LAL LAL (Station ID 232210)

Example data for water quality data for Black River at South Forest, Tasmania (Station ID 314213)
}
\seealso{
\code{\link{dataBassRiver}}

\code{\link{dataBassRiver}}
}
\keyword{datasets}
