% Generated by roxygen2: do not edit by hand
\docType{data}
\name{data_P_WL}
\alias{data_P_WL}
\title{Example sub-daily rainfall and tidal water level data}
\format{
A list with streamflow and catchment average precipitation and temperature for the following sites: \cr
120301B, 602004, 235203, 410044, 105105A, \cr
corresponding to Arid, Mediterranean, Temperate, Subtropical, and Tropical climates. \cr
Catchment areas are 35326, 2433, 721, 1072, 297 km2 respectively. \cr
Streamflow is from the Australian Bureau of Meteorology Hydrologic Reference Station network \cr
and catchment average climate variables were extracted using AWAPer.
}
\usage{
data_P_WL
}
\description{
Hourly rainfall (P) and water level (WL) at Burnie, Tasmania for 1997-01-14 to 1997-02-14 \cr
(Pluvio ID: 091009; Tide gauge: IDO71005)
}

\source{
Sub-daily rainfall data are from Australian Bureau of Meteorolgy: \cr
\url{http://www.bom.gov.au/climate/data/stations/} \cr
Sub-daily tidal water level data are from Australian Bureau of Meteorolgy \cr
Australian Baseline Sea Level Monitoring Project: \cr
\url{http://www.bom.gov.au/oceanography/projects/abslmp/data/index.shtml}
}
\keyword{datasets}
