% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dataBassRiver}
\alias{dataBassRiver}
\title{Streamflow data}
\format{
A vector of 67 daily streamflow values in (ML/day)
}
\usage{
dataBassRiver
}
\description{
Streamflow data for Bass River at Loch (227219A) for 30/06/1974-04/09/1974
}
\details{
This data is obtained from Grayson et al (1996)
}
\references{
Grayson, R., Argent, R. M., Nathan, R. J., McMahon, T. A. & Mein, R. G. (1996) Hydrological
Recipes, Cooperative Research Centre for Catchment Hydrology, Melbourne.
}
\seealso{
\code{\link{dataLoch}}
}
\keyword{datasets}
