% Generated by roxygen2: do not edit by hand
\docType{data}
\name{dataCatchment}
\alias{dataCatchment}
\title{Catchment data}
\format{
A list with streamflow and catchment average precipitation and temperature for the following sites: \cr
120301B, 602004, 235203, 410044, 105105A, \cr
corresponding to Arid, Mediterranean, Temperate, Subtropical, and Tropical climates. \cr
Catchment areas are 35326, 2433, 721, 1072, 297 km2 respectively. \cr
Streamflow is from the Australian Bureau of Meteorology Hydrologic Reference Station network \cr
and catchment average climate variables were extracted using AWAPer.
}
\usage{
dataCatchment
}
\description{
Example data for five sites across Australia
}
\source{
\url{http://www.bom.gov.au/water/hrs/}
}
\references{
Jones, D., Wang, W., & Fawcett, R., 2009. High-quality spatial climate data-sets for \cr
Australia. Aust. Meteorol. Oceanogr. J. 58, 233-248.
}
\keyword{datasets}
