% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dataHRS_sample}
\alias{dataHRS_sample}
\title{Catchment data}
\format{
A data frame with 1559 rows and 5 variables:
Each element contains data for a catchment in the order of:
A5030502, 314213, G8150018, G0050115, 410061; with corresponding names:
Scott Creek, Black River, Elizabeth River, Hugh River and Adelong Creek
Each of the 5 elements contains 4 data variables:
$P, $E, $Q: the time-series of daily P, PET and Q, all in mm/d.
Q was converted from gauge values (ML/d) to mm/d by dividing catchment area, P and PET are averaged over catchment area.
$site: the HRS gauge ID
}
\source{
\url{http://somewhere.important.com/}
}
\usage{
dataHRS_sample
}
\description{
Example data for Bass River at Loch
}
\examples{
plot(NULL, xlim = c(110, 160), ylim = c(-50, 0), type = "n")
for (i in 1:length(mapAust50m)) {
lines(mapAust50m[[i]])
}
for (j in 1:5) {
 dataHRS_sample[[i]]
}
}
\keyword{datasets}
