% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downloadHRS.R
\name{downloadHRS}
\alias{downloadHRS}
\title{Download HRS data}
\usage{
downloadHRS(setpath = getwd(), setpath.meta = setpath)
}
\arguments{
\item{setpath}{Directory data will be downloaded}
}
\description{
This function removes baseflow based on a recursive digital filter and is based on
the implementation in the EcoHydRology package with initial values taken from Fuka et al. (2018).

The formulation is after Lyne and Hollick (1979) and nicely described in Furey and Gupta (2001).
Reccomended parameters are after Nathan and McMahon (1990). Additional discussion on implementations
can be found in Ladson et al (2013).
}
\examples{
downloadHRS()
}
\references{
Furey, P., & Gupta, V. (2001). A Physically Based Filter for Spearating Base
Flow from Streamflow Time Series. Water Resources Research, 37(11), 2709–2722.

Ladson, A., Brown, R., Neal, B., & Nathan, R. (2013). A standard approach to baseflow separation using the Lyne and Hollick filter.
Australian Journal of Water Resources, 17(1).

Nathan, R. J., & McMahon, T. A. (1990). Evaluation of automated techniques for base flow and recession analyses.
Water Resources Research, 26(7), 1465–1473.
}
\keyword{baseflow}
