% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hourlyQ}
\alias{hourlyQ}
\title{Hourly streamflow data}
\format{
A data frame with two columns: time stamp (hourly) and streamflow (units?)
}
\source{
\url{http://www.bom.gov.au/water/hrs/ check??}
}
\usage{
dataCatchment
}
\description{
Hourly streamflow data from site q138903A. check??
}
\keyword{datasets}
