% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mapAust50m}
\alias{mapAust50m}
\title{Australia map}
\format{
A list of 41 matrices of x, y values for plotting lines of Australia's coast
}
\source{
\url{https://www.naturalearthdata.com/downloads/50m-physical-vectors/}
}
\usage{
mapAust50m
}
\description{
Australia's coastline at a resolution of 1:50m
}
\details{
The projection string is: "CRS arguments: +proj=longlat +datum=WGS84 +no_defs"
}
\examples{
plot(NULL, xlim = c(110, 160), ylim = c(-50, 0), type = "n", xlab = "", ylab = "")
for (i in 1:length(mapAust50m)) {
  lines(mapAust50m[[i]])
}
}
\keyword{datasets}
