% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pairEvents.R
\name{pairEvents}
\alias{pairEvents}
\title{Pair rainfall to streamflow events}
\usage{
pairEvents(events.1, events.2, lag = 5, type = 1)
}
\arguments{
\item{lag}{Maximum lag time (search radius) for pairing}

\item{srt.1}{Start time of the dependent events used for pairing}

\item{end.1}{End time of the dependent events used for pairing}

\item{srt.2}{Start time of the independent events used for pairing}

\item{end.2}{End time of the independent events used for pairing}
}
\value{
Returns indices of start and end of events as a two column dataframe.
}
\description{
Pairing can be performed on the basis of rainfall or streamflow events within a given lag time.
}
\details{
Pairing can be performed on the basis of rainfall or streamflow events within a given lag time.
It is appropriate to pick a lag time that is equiavelent to the catchment time of concentration.
}
\examples{
## STREAMFLOW EXAMPLES
library(hydroEvents)
srt = as.Date("2011-05-01")
end = as.Date("2011-08-30")
data = data314213[which(data314213$Date >= srt & data314213$Date <= end),]
events.P = eventPOT(data$precip_mm, threshold = 1, min.diff = 2)
bf = baseFlow(data$Flow_ML, alpha = 0.925)
events.Q = eventMaxima(data$Flow_ML-bf, delta.y = 1, delta.x = 1, thresh = 0) # type 1
events.Q = eventMaxima(data$Flow_ML-bf, delta.y = 1, delta.x = 1, thresh = 10) # type 2
events.Q = eventMaxima(data$Flow_ML-bf, delta.y = 500, delta.x = 1, thresh = 100) # type 3
events.Q = eventMaxima(data$Flow_ML-bf, delta.y = 500, delta.x = 7, thresh = 100) # type 3
print(events.Q)
#events.Q = eventBaseflow(data$Flow_ML-bf)
par(mfrow = c(2, 1), mar = c(2, 4, 4, 4))
plotEvents(data = pmax(data$precip_mm - 1, 0), events = events.P, type = "hyet")
plotEvents(data = data$Flow_ML-bf, events = events.Q, type = "lineover")
dev.off()
matched = pairEvents(events.P, events.Q, lag = 5, type = 1)   # thres 0
matched = pairEvents(events.P, events.Q, lag = 10, type = 2)  # thres 10
matched = pairEvents(events.P, events.Q, lag = 5, type = 3)  # thres 0
matched = pairEvents(events.P, events.Q, lag = 10, type = 4)  # thres 0
print(matched)
plotPairedEvents(data.1 = pmax(data$precip_mm), data.2 = data$Flow_ML-bf, events = matched, type = "hyet", col = rainbow)
events.1 = events.P
events.2 = events.Q
lag = 10
type = 3
}
\references{
None
}
\keyword{baseflow}
\keyword{events}
