% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smoothing.R
\name{smoothing}
\alias{smoothing}
\title{Perform MA smoothing on data}
\usage{
smoothing(data, nstep = 1, med.weight = 1)
}
\arguments{
\item{srt}{Vector of indices for the event start}

\item{end}{Vector of indices for the event end}

\item{f.list}{List of functions to be applied to the events}
}
\value{
Returns a dataframe where the row is each event and the column is each statistic
}
\description{
To fix
}
\details{
No additional details provided.
}
\examples{
# Extract event statistics from quickflow
bf = baseFlow(bassRiver, alpha = 0.925)
qf = bassRiver - bf
event.indices = eventPOT(qf)
event.stats = eventStats(event.indices$srt, event.indices$end, qf)

# Plot maxima of events
plot(1:length(bassRiver), qf, type = "h", lwd = 2, col = "steelblue",
     ylab = "Flow (ML/d)", xlab = "Time index", mgp = c(2, 0.6, 0))
points(event.indices$srt + event.stats$which.max - 1, event.stats$max, col = "red", pch = 16, cex = 1.2)
legend("topright", legend = c("Quickflow", "Max"), cex = 0.8,
       lwd = c(2, NA), pch = c(NA, 16), col = c("steelblue", "red"), bty = "n")
}
\references{
\url{https://onlinelibrary.wiley.com/doi/epdf/10.1002/hyp.11185}
}
\seealso{
\code{\link{eventPOT}}
}
\keyword{events}
