% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data_P_WL}
\alias{data_P_WL}
\title{Example sub-daily rainfall and tidal water level data}
\format{
Each of P and WL data is a simple vector with no timestamp. The original data is in hourly time step.
}
\source{
Sub-daily rainfall data are from Australian Bureau of Meteorolgy:
\url{http://www.bom.gov.au/climate/data/stations/}
Sub-daily tidal water level data are from Australian Bureau of Meteorolgy Australian Baseline Sea Level Monitoring Project:
\url{http://www.bom.gov.au/oceanography/projects/abslmp/data/index.shtml}
}
\usage{
data_P_WL
}
\description{
Hourly rainfall (P) and water level (WL) at Burnie, Tasmania for 1997-01-14 to 1997-02-14 (Pluvio ID: 091009; Tide gauge: IDO71005)
}
\keyword{datasets}
