% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dataCatchment}
\alias{dataCatchment}
\title{Catchment data}
\format{
A list with streamflow and catchment average precipitation and temperature for the following sites:
120301B, 602004, 235203, 410044, 105105A, corresponding to Arid, Mediterranean, Temperate, Subtropical, and Tropical climates.
Catchment areas are 35326, 2433, 721, 1072, 297 km2 respectively.
Streamflow is from the Australian Bureau of Meteorology Hydrologic Reference Station network and catchment average climate variables
were extracted using AWAPer.
}
\source{
\url{http://www.bom.gov.au/water/hrs/}
}
\usage{
dataCatchment
}
\description{
Example data for five sites across Australia
}
\references{
Peterson, T.J., Wasko, C., Saft., & Peel, M.C. (2020) AWAPer: An R package for area weighted catchment
daily meteorological data anywhere within Australia, Hydrological Processes, 34, 1301-1306.

Jones, D., Wang, W., & Fawcett, R., 2009. High-quality spatial climate data-sets for Australia. Aust. Meteorol. Oceanogr. J. 58, 233–248.
}
\keyword{datasets}
