% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dataLoch}
\alias{dataLoch}
\title{Rainfall data}
\format{
A vector of 67 daily rainfall values in (mm)
}
\source{
\url{http://www.bom.gov.au/climate/data/stations/}
}
\usage{
dataLoch
}
\description{
Rainfall data for Loch (Station ID 086067) for 30/06/1974-04/09/1974
}
\seealso{
\code{\link{dataBassRiver}}
}
\keyword{datasets}
