% Generated by roxygen2: do not edit by hand

\name{eventBaseflow}
\alias{eventBaseflow}
\title{Event identification (using baseflow index)}
\usage{
eventBaseflow(
  data,
  BFI_Th = 0.5,
  min.diff = 1,
  threshold = 0,
  out.style = "summary"
)
}
\arguments{
\item{data}{The data vector (e.g. a streamflow time series)}

\item{BFI_Th}{Minimum BFI to identify baseflow}

\item{min.diff}{Minimum length for an event}

\item{threshold}{Value above which an event is considered to have occurred}

\item{out.style}{The type of output (currently either "summary" or "none")}
}
\value{
Returns indices of start and end of events as a two column dataframe and event statistics
}
\description{
Events are identified on the basis BFI etc etc
}
\details{
filter.type can be a "simple" weigthed moving average or smoothing "spline".
If \code{delta.y} is negative it is applied a fractional decrease from the peak
}
\examples{
data(dataBassRiver)
bf = baseflowB(dataBassRiver)
BFI_res = eventBaseflow(dataBassRiver, BFI_Th = 0.5, min.diff = 1, threshold = 0)

}
\seealso{
\code{\link{calcStats}} \code{\link{eventPOT}}
}
\keyword{events}
