% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eventMinima.R
\name{eventMinima}
\alias{eventMinima}
\title{Event identification (using local minima as a basis)}
\usage{
eventMinima(
  data,
  delta.x = 5,
  delta.y = 20,
  thresh = -1,
  out.style = "summary"
)
}
\arguments{
\item{data}{The data vector (e.g. a streamflow time series)}

\item{delta.x}{Minimum length for an event}

\item{delta.y}{Maximum difference in data between start and end of an event}

\item{filter.type}{c("simple", "spline") Optional smoothing of data series}
}
\value{
Returns indices of start and end of events as a two column dataframe and event statistics
}
\description{
Events are identified on the basis of local minima with an "event" considered to have
occurred once the data has returned to within a threshold level of the start of the event.
}
\details{
filter.type can be a "simple" weigthed moving average or smoothing "spline"
}
\examples{
# Example extracting events from quickflow
bf = baseFlow(dataBassRiver, alpha = 0.925)
qf = dataBassRiver - bf
events = eventMinima(qf, delta.x = 5, delta.y = 20)
# delta.x = 5, delta.y = 20 # 5 events identified
# delta.x = 5, delta.y = 10 # 4 events identified
# delta.x = 1, delta.y = 20 # 6 events identified
plot(1:length(qf), qf, type = "l", lwd = 1, col = "black", main = "Events with maxima identified",
  ylab = "Quickflow (ML/day)", xlab = "Time index", mgp = c(2, 0.6, 0))
n.events = nrow(events)
for (i in 1:n.events) {
  idx = events$srt[i]:events$end[i]
  lines(idx, qf[idx], col = rainbow(nrow(events))[i], lwd = 2)
}
points(events$srt + events$which.max - 1, qf[events$srt + events$which.max - 1], cex = 1.2, lwd = 2)
}
\references{
Tang and Carey (2017) HydRun: A MATLAB toolbox for rainfall-runoff analysis,
Hydrological Processes (31) 2670-2682
}
\seealso{
\code{\link{calcStats}} \code{\link{eventPOT}}
}
\keyword{events}
