% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/limbs.R
\name{limbs}
\alias{limbs}
\title{Extract rising/falling limbs}
\usage{
limbs(
  data,
  dates = NULL,
  events,
  plot = T,
  main = "Event hydrographs",
  filter = F,
  min.rates = c(0, 0)
)
}
\arguments{
\item{data}{The data vector (e.g. a streamflow time series)}

\item{delta.y}{Minimum allowable difference from a peak to a trough}

\item{delta.x}{Minimum length for an event}

\item{filter.type}{c("simple", "spline") Optional smoothing of data series}

\item{...}{Further arguments passed eventStats}
}
\value{
Returns indices of start and end of events as a two column dataframe and event statistics
}
\description{
...
}
\details{
filter.type can be a "simple" weigthed moving average or smoothing "spline".
If \code{delta.y} is negative it is applied a fractional decrease from the peak
}
\examples{
# Example extracting events from quickflow
bf = baseFlow(dataBassRiver, alpha = 0.925)
qf = dataBassRiver - bf
events = eventMaxima(qf, delta.y = 200, delta.x = 1) # 5 events identified
# delta.y = 200, delta.x = 1 # 5 events identified
# delta.y = 500, delta.x = 1 # 3 events identified
# delta.y = 10,  delta.x = 7 # 2 events identified
plot(1:length(qf), qf, type = "l", lwd = 1, col = "black", main = "Events with maxima identified",
     ylab = "Quickflow (ML/day)", xlab = "Time index", mgp = c(2, 0.6, 0))
n.events = nrow(events)
for (i in 1:n.events) {
  idx = events$srt[i]:events$end[i]
  lines(idx, qf[idx], col = rainbow(nrow(events))[i], lwd = 2)
}
points(events$srt + events$which.max - 1, qf[events$srt + events$which.max - 1], cex = 1.2, lwd = 2)
}
\references{
...
}
\seealso{
\code{\link{eventStats}} \code{\link{eventPOT}}
}
\keyword{events}
