% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{WQ_Q}
\alias{WQ_Q}
\title{Example water quality and streamflow data}
\format{
Water quality (WQ) and streamflow (Q) data at matching time steps from 4 HRS catchments.
Each dataset (qdata and wqdata) is a list of length 4, corresponding to the 4 catchments.
}
\source{
HRS streamflow data:
\url{http://www.bom.gov.au/water/hrs/}
water quality data:
WaterNSW: \url{https://www.waternsw.com.au/waterinsights/real-time-data}
Northern Territory Department of Environment, Parks and Water Security \url{https://water.nt.gov.au/}
South Australia Department for Environment and Water \url{https://www.waterconnect.sa.gov.au/}
}
\usage{
WQ_Q
}
\description{
Data from 4 HRS (Hydrologic Reference Stations, Australian Bureau of Meteorolgy) catchments are included:
catchment IDs: 410073, 424002, G8150018, A5020502.
}
\keyword{datasets}
